requirements.txt
addict
diffusers
gradio
huggingface_hub
matplotlib
numpy==1.26.4
onnxruntime
opencv_python
Pillow
pycocotools
PyYAML
requests
setuptools
supervision
termcolor
timm
torch
torchvision
transformers
yapf
nltk
fairscale
litellm
